refactor(ui): split (landing) vs (detail) route groups for <MatchStrip>
Consolidate the per-page <MatchStripSlot /> mounts introduced in #1641 into
a shared (landing) route-group layout, matching the locked Phase 3 spec at
docs/design/mockups/phase-3-c-header-and-matchstrip/matchstrip-locked.md.

- Add app/(landing)/layout.tsx — mounts <MatchStripSlot /> once before children.
- Move app/page.tsx + four landing index pages and their co-located helpers
  into app/(landing)/{nieuws,events,jeugd,sponsors}/. Detail [slug] subdirs
  stay in (main).
- Remove the five inline <MatchStripSlot /> mounts and their imports.
- Promote loading-envelope.test.tsx to app/__tests__/ and extend the
  completeness glob to scan both (main) and (landing) route groups.
- Update stale (main)/* import paths in stories and CLAUDE.md.

URL paths are unchanged (Next.js collapses route-group parens). Build,
type-check, and the full vitest suite pass.

Closes #1662
