+++ apps/web/src/app/(landing)/events/loading.tsx
