+++ apps/web/src/app/(main)/layout.tsx
@@ -3,7 +3,7 @@
  * absent here per Phase 3 Checkpoint C spec
  * (`docs/design/mockups/phase-3-c-header-and-matchstrip/matchstrip-locked.md`).
  * Landing pages (homepage + section indexes) live under (landing) which mounts
- * `<MatchStrip />` in its own layout.
+ * `<MatchStripSlot />` in its own layout.
  */
 export default function MainLayout({
   children,
