+++ apps/web/src/components/sponsors/SponsorsPage/SponsorsPage.stories.tsx
@@ -9,7 +9,7 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import { SponsorsPage } from "./SponsorsPage";
 import { mockSponsors } from "../Sponsors.mocks";
-import SponsorsLoading from "@/app/(main)/sponsors/loading";
+import SponsorsLoading from "@/app/(landing)/sponsors/loading";
 
 // ---------------------------------------------------------------------------
 // Meta
